+++ src/GUI/Main.java
@@ -11,13 +11,7 @@ public class Main {
 
     public static void main(String[] args) {
         Menu inicio = new MenuPrincipal();
-        // inicio.mostrar();
+        inicio.mostrar();
         Conexion conexion = Conexion.getInstance();
-
-        UsuarioController uc = new UsuarioController();
-        List<Usuario> usuarios = uc.listarUsuarios();
-        for (Usuario u : usuarios) {
-            System.out.println(u);
-}
     }
 }